requirements.txt
asyncio==3.4.3
websockets==12.0
pandas==2.2.1
python-dotenv==1.0.1
